Move `SortByRunState()` into `RepairRunService`.
