+++ src/server/src/main/java/io/cassandrareaper/core/RepairRun.java
@@ -18,8 +18,6 @@
 package io.cassandrareaper.core;
 
 import java.util.Collections;
-import java.util.Comparator;
-import java.util.List;
 import java.util.Objects;
 import java.util.Set;
 import java.util.UUID;
@@ -347,19 +345,4 @@ public final class RepairRun implements Comparable<RepairRun> {
     }
   }
 
-  public static void sortByRunState(List<RepairRun> repairRunCollection) {
-    Comparator<RepairRun> comparator = new Comparator<RepairRun>() {
-      @Override
-      public int compare(RepairRun o1, RepairRun o2) {
-        if (!o1.getRunState().isTerminated() && o2.getRunState().isTerminated()) {
-          return -1; // o1 appears first.
-        }  else if (o1.getRunState().isTerminated() && !o2.getRunState().isTerminated()) {
-          return 1; // o2 appears first.
-        } else { // Both RunStates have equal isFinished() values; compare on time instead.
-          return o1.getId().compareTo(o2.getId());
-        }
-      }
-    };
-    Collections.sort(repairRunCollection, comparator);
-  }
 }
