+++ src/server/src/main/java/io/cassandrareaper/service/RepairRunService.java
@@ -35,6 +35,7 @@ import java.util.Arrays;
 
 import java.util.Collection;
 import java.util.Collections;
+import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
@@ -83,6 +84,22 @@ public final class RepairRunService {
 
   public static RepairRunService create(AppContext context) {
     return new RepairRunService(context, () -> ClusterFacade.create(context));
+  }
+
+  public static void sortByRunState(List<RepairRun> repairRunCollection) {
+    Comparator<RepairRun> comparator = new Comparator<RepairRun>() {
+      @Override
+      public int compare(RepairRun o1, RepairRun o2) {
+        if (!o1.getRunState().isTerminated() && o2.getRunState().isTerminated()) {
+          return 1; // o2 appears first.
+        }  else if (o1.getRunState().isTerminated() && !o2.getRunState().isTerminated()) {
+          return -1; // o1 appears first.
+        } else { // Both RunStates have equal isFinished() values; compare on time instead.
+          return o1.getId().compareTo(o2.getId());
+        }
+      }
+    };
+    Collections.sort(repairRunCollection, comparator);
   }
 
   /**
